Cryptocurrency trading-pair forecasting, using machine learning and deep learning Technique
The goal of this project is to help the individual crypto-currency investor that might not have the needed resource to compete in the fast-paced financial market. We intend to do this by building a model that would help an investor in cryptocurrency pick suitable pair cryptocurrencies for pair trade, and help the forecast the price movement of the currencies and suggest a possible trading strategy that could earn the trader/ investor some profitable trade.
Our objectives are as follows
•	Find suitable pair for pair cryptocurrency trade
•	Predict cryptocurrency price movement
•	Develop and suggest a pair trading strategy